fast-api-sql-demo/requirements.txt
# fast api basic
fastapi~=0.109.2
uvicorn[standard]~=0.27.1
# test
pytest~=8.1.1
pytest-asyncio~=0.23.6
httpx~=0.27.0
# seems to be included in fast api might be not relvant
pydantic~=2.6.2
# sql database
SQLAlchemy~=2.0.27
psycopg2-binary~=2.9.9
alembic~=1.13.1
# auth
passlib[bcrypt]~=1.7.4
bcrypt~=4.1.2
python-multipart~=0.0.9
python-jose[cryptography]~=3.3.0
